+++ sdk/storage/Azure.ResourceManager.Storage/src/Customize/Models/StorageActiveDirectoryProperties.cs
@@ -5,9 +5,13 @@
 
 using System;
 using System.ComponentModel;
+using System.Runtime.CompilerServices;
+using System.Text.Json;
+using Azure.Core;
 
 namespace Azure.ResourceManager.Storage.Models
 {
+    [CodeGenSerialization(nameof(ActiveDirectoryDomainGuid), DeserializationValueHook = nameof(DeserializeNullableGuid))]
     public partial class StorageActiveDirectoryProperties
     {
         /// <summary> Initializes a new instance of <see cref="StorageActiveDirectoryProperties"/>. </summary>
@@ -32,6 +36,19 @@ namespace Azure.ResourceManager.Storage.Models
         {
             get => ActiveDirectoryDomainGuid ?? Guid.Empty;
             set => ActiveDirectoryDomainGuid = value == Guid.Empty ? null : value;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static void DeserializeNullableGuid(JsonProperty property, ref Guid? domainGuid)
+        {
+            if (string.IsNullOrEmpty(property.Value.GetString()))
+            {
+                domainGuid = null;
+            }
+            else
+            {
+                domainGuid = property.Value.GetGuid();
+            }
         }
     }
 }